fixed multiple pdf files; add report type based output
